+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fix create role and create group Admin APIs [#386](https://github.com/rokwire/core-building-block/issues/386)
 
+### Changed
+- Return ordered list of account auth types on link account auth type [#376](https://github.com/rokwire/core-building-block/issues/376)
+
 ## [1.18.0] - 2022-01-25
 ### Added
 - Expose create app org group admin API [#309](https://github.com/rokwire/core-building-block/issues/309)